+++ CHANGELOG.md
@@ -1,7 +1,17 @@
+<!--
+ * @Author: atdow
+ * @Date: 2017-08-21 14:59:59
+ * @LastEditors: null
+ * @LastEditTime: 2022-10-29 17:55:50
+ * @Description: file description
+-->
+
 # Change Log
-All notable changes to the "vue-peek" extension will be documented in this file.
+
+All notable changes to the "vue-jumper" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
-- Initial release
+
+* Initial release